requirements.txt
# 代理记账客户管理系统依赖
# Accounting Customer Management System Dependencies

# Flask 核心
Flask==2.3.3
Werkzeug==2.3.7

# 数据库
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21

# 多语言支持
Flask-Babel==4.0.0
Babel==2.12.1

# 表单处理
Flask-WTF==1.1.1
WTForms==3.0.1

# 跨域支持
Flask-CORS==4.0.0

# JWT 认证
PyJWT==2.8.0
Flask-JWT-Extended==4.5.3

# 环境变量管理
python-dotenv==1.0.0

# 日期时间处理
python-dateutil==2.8.2
pytz==2023.3

# 数据验证
email-validator==2.0.0

# 开发工具
pytest==7.4.2
pytest-flask==1.2.0

# 生产环境
gunicorn==21.2.0
